got some PEP fixes and now only creater can edit/delete contents.

+++ app.py
@@ -1,12 +1,17 @@
-from flask import Flask, render_template, request, url_for, redirect, flash, jsonify, make_response
+from flask import Flask, render_template, request, url_for
+from flask import redirect, flash, jsonify, make_response
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, joinedload
 from database_setup import Category, CategoryItem, Base
 from flask import session as login_session
-import random, string
+import random
+import string
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
-import httplib2, json, requests
+import httplib2
+import json
+import requests
+
 
 app = Flask(__name__)
 
@@ -16,105 +21,170 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
-CLIENT_ID = json.loads(open('client_secret.json', 'r').read())['web']['client_id']
+CLIENT_ID = json.loads(open(
+                'client_secret.json', 'r').read())['web']['client_id']
 
 
 # Index page
 @app.route('/')
 def showAll():
-    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+    state = ''.join(random.choice(
+                string.ascii_uppercase + string.digits) for x in xrange(32))
     login_session['state'] = state
     title = "Welcome to Item Catalog Website"
     categories = session.query(Category).all()
-    recentItems = session.query(CategoryItem).order_by(CategoryItem.datetime.desc()).limit(10)
-    return render_template('index.html', title=title, categories=categories, recentItems=recentItems, STATE=state, login_session=login_session)
+    recentItems = session.query(CategoryItem).order_by(
+                    CategoryItem.datetime.desc()).limit(10)
+    return render_template(
+                'index.html',
+                title=title, categories=categories,
+                recentItems=recentItems, STATE=state,
+                login_session=login_session)
+
 
 # show category and it's items
 @app.route('/category/<int:category_id>')
 def showCategoryItems(category_id):
-    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+    state = ''.join(random.choice(
+                string.ascii_uppercase + string.digits) for x in xrange(32))
     login_session['state'] = state
     categories = session.query(Category).all()
     cur_category = session.query(Category).filter_by(id=category_id).one()
     title = "Items of {}".format(cur_category.name.capitalize())
-    items = session.query(CategoryItem).filter_by(category_id=category_id).order_by(CategoryItem.datetime.desc()).all()
-    return render_template('category.html', title=title, categories=categories, cur_category=cur_category, items=items, STATE=state, login_session=login_session)
+    items = session.query(CategoryItem).filter_by(
+                category_id=category_id
+            ).order_by(CategoryItem.datetime.desc()).all()
+    return render_template(
+            'category.html',
+            title=title, categories=categories,
+            cur_category=cur_category, items=items,
+            STATE=state, login_session=login_session)
+
 
 # show item's detail
 @app.route('/item/<int:item_id>')
 def showItemDetails(item_id):
-    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+    state = ''.join(random.choice(
+                string.ascii_uppercase + string.digits) for x in xrange(32))
     login_session['state'] = state
     categories = session.query(Category).all()
     cur_item = session.query(CategoryItem).filter_by(id=item_id).one()
     title = "Details of {}".format(cur_item.name.capitalize())
-    return render_template('item.html', title=title, cur_item=cur_item, STATE=state, login_session=login_session)
+    return render_template(
+            'item.html',
+            title=title, cur_item=cur_item,
+            STATE=state, login_session=login_session)
+
 
 # add a new category
 @app.route('/category/new', methods=['GET', 'POST'])
 def addCategory():
+    if "username" not in login_session:
+        return redirect("/")
     if request.method == 'POST':  # save a new category
-        newCategory = Category(name=request.form['categoryName'])
+        newCategory = Category(
+                        name=request.form['categoryName'],
+                        creater=login_session["email"])
         session.add(newCategory)
         session.commit()
         flash('New category created!!')
         return redirect('/')
     else:
-        state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+        state = ''.join(random.choice(
+                    string.ascii_uppercase + string.digits
+                    ) for x in xrange(32))
         login_session['state'] = state
         title = "Add a new category"
-    return render_template('addcategory.html', title=title, login_session=login_session)
+    return render_template(
+            'addcategory.html',
+            title=title, login_session=login_session)
+
 
 # edit a category
 @app.route('/category/edit/<int:category_id>', methods=['GET', 'POST'])
 def editCategory(category_id):
+    if "username" not in login_session:
+        return redirect("/")
     if request.method == 'POST':  # save a new category
-        targetCategory = session.query(Category).filter_by(id=category_id).one()
+        targetCategory = session.query(Category).filter_by(
+                                                    id=category_id).one()
+        if targetCategory.creater != login_session['email']:
+            flash('You can only edit the categories you created.')
+            return redirect("/")
         targetCategory.name = request.form['categoryName']
         session.add(targetCategory)
         session.commit()
         flash('Category edited successfully!!')
         return redirect('/')
     else:
-        targetCategory = session.query(Category).filter_by(id=category_id).one()
-        state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+        targetCategory = session.query(Category).filter_by(
+                                                id=category_id).one()
+        state = ''.join(random.choice(
+                    string.ascii_uppercase + string.digits
+                    ) for x in xrange(32))
         login_session['state'] = state
         title = "Edit {} category".format(targetCategory.name)
-    return render_template('editcategory.html', title=title, login_session=login_session, targetCategory=targetCategory)
+    return render_template(
+            'editcategory.html',
+            title=title,
+            login_session=login_session,
+            targetCategory=targetCategory)
+
 
 # delete a category
 @app.route('/category/delete/<int:category_id>', methods=['GET'])
 def deleteCategory(category_id):
+    if "username" not in login_session:
+        return redirect("/")
     targetCategory = session.query(Category).filter_by(id=category_id).one()
+    if targetCategory.creater != login_session['email']:
+        flash('You can only delete the categories you created.')
+        return redirect("/")
     session.delete(targetCategory)
     session.commit()
     # todo : need to delete all belonging items as well
     flash('Category deleted successfully!!')
     return redirect('/')
 
+
 # add a new item
 @app.route('/item/new', methods=['GET', 'POST'])
 def addItem():
+    if "username" not in login_session:
+        return redirect("/")
     if request.method == 'POST':  # save a new category
-        newItem = CategoryItem(name=request.form['itemname'],
+        newItem = CategoryItem(
+                    name=request.form['itemname'],
                     category_id=request.form['category_id'],
-                    description=request.form['description'])
+                    description=request.form['description'],
+                    creater=login_session["email"])
         session.add(newItem)
         session.commit()
         flash('New item created!!')
         return redirect('/')
     else:
-        state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+        state = ''.join(random.choice(
+                    string.ascii_uppercase + string.digits
+                    ) for x in xrange(32))
         login_session['state'] = state
         title = "Add a new item"
         categories = session.query(Category).all()
-    return render_template('additem.html', title=title, login_session=login_session, categories=categories)
+    return render_template(
+            'additem.html',
+            title=title, login_session=login_session,
+            categories=categories)
+
 
 # edit an item
 @app.route('/item/edit/<int:item_id>', methods=['GET', 'POST'])
 def editItem(item_id):
+    if "username" not in login_session:
+        return redirect("/")
     if request.method == 'POST':  # save a edited item
         targetItem = session.query(CategoryItem).filter_by(id=item_id).one()
+        if targetItem.creater != login_session['email']:
+            flash('You can only edit the items you created.')
+            return redirect("/")
         targetItem.name = request.form['itemname']
         targetItem.description = request.form['description']
         targetItem.category_id = request.form['category_id']
@@ -123,27 +193,44 @@ def editItem(item_id):
         flash('Item edited successfully!!')
         return redirect('/')
     else:
-        state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+        state = ''.join(random.choice(
+                    string.ascii_uppercase + string.digits
+                    ) for x in xrange(32))
         login_session['state'] = state
         title = "Add a new item"
         categories = session.query(Category).all()
         targetItem = session.query(CategoryItem).filter_by(id=item_id).one()
-    return render_template('edititem.html', title=title, login_session=login_session, categories=categories, targetItem=targetItem)
+    return render_template(
+            'edititem.html',
+            title=title, login_session=login_session,
+            categories=categories, targetItem=targetItem)
+
 
 # delete an item
 @app.route('/item/delete/<int:item_id>', methods=['GET'])
 def deleteItem(item_id):
+    if "username" not in login_session:
+        return redirect("/")
     targetItem = session.query(CategoryItem).filter_by(id=item_id).one()
+    if targetItem.creater != login_session['email']:
+        flash('You can only delete the items you created.')
+        return redirect("/")
     session.delete(targetItem)
     session.commit()
     flash('Item deleted successfully!!')
     return redirect('/')
 
+
 # JSON API - whole catalog
 @app.route('/catalog.json')
 def categoryJSON():
-    categories = session.query(Category).options(joinedload(Category.items)).all()
-    return jsonify(category=[dict(c.serialize, items=[i.serialize for i in c.items]) for c in categories])
+    categories = session.query(Category).options(
+                    joinedload(Category.items)).all()
+    return jsonify(
+            category=[
+                dict(c.serialize, items=[i.serialize for i in c.items])
+                for c in categories])
+
 
 # JSON API - all categories
 @app.route('/categories/JSON')
@@ -151,17 +238,20 @@ def categoriesJSON():
     categories = session.query(Category).all()
     return jsonify(category=[c.serialize for c in categories])
 
+
 # JSON API - all items
 @app.route('/items/JSON')
 def itemsJSON():
     items = session.query(CategoryItem).all()
     return jsonify(items=[i.serialize for i in items])
 
+
 # JSON API - details of an item
 @app.route('/item/<int:item_id>/JSON')
 def itemDeatilJSON(item_id):
     item = session.query(CategoryItem).filter_by(id=item_id).one()
     return jsonify(item=[item.serialize])
+
 
 # Google login connect
 @app.route('/gconnect', methods=['POST'])
@@ -216,8 +306,8 @@ def gconnect():
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
-                                 200)
+        response = make_response(
+                    json.dumps('Current user is already connected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -235,8 +325,10 @@ def gconnect():
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['gplus_id'] = data['id']
+    login_session['email'] = data['email']
 
     return 'success!'
+
 
 # Google login disconnect
 @app.route('/gdisconnect')
@@ -244,13 +336,15 @@ def gdisconnect():
     access_token = login_session.get('access_token')
     if access_token is None:
         print 'Access Token is None'
-        response = make_response(json.dumps('Current user not connected.'), 401)
+        response = make_response(
+                    json.dumps('Current user not connected.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
     print 'In gdisconnect access token is %s', access_token
     print 'User name is: '
     print login_session['username']
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = 'https://accounts.google.com/\
+            o/oauth2/revoke?token=%s' % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is '
@@ -260,12 +354,14 @@ def gdisconnect():
         del login_session['gplus_id']
         del login_session['username']
         del login_session['picture']
-        # response = make_response(json.dumps('Successfully disconnected.'), 200)
+        # response = make_response(
+        #               son.dumps('Successfully disconnected.'), 200)
         # response.headers['Content-Type'] = 'application/json'
         # return response
         return redirect('/')
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(
+                    json.dumps('Failed to revoke token for given user.', 400))
         response.headers['Content-Type'] = 'application/json'
         return response
